Add net-libs/webkit-gtk that merges on Raspberry Pi
The append-ldflags part is untested, as due to the compile times after a failure
had to add it manually. Ought to work though, let me know if not.
USE=rpi is used to pick up rpi-userland gles2/egl together with USE=webgl, which
actually also enables accelerated compositing..
As newer version does not support building without accelerated compositing, and
with that a GL implementation will be mandatory, no point in introducing new USE
flags in an overlay to separate webgl and AC. metadata.xml is updated here, though